requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-telegram-bot==21.3
anthropic==0.28.0
supabase==2.5.0
boto3==1.34.131
python-dotenv==1.0.1
jinja2==3.1.4
python-multipart==0.0.9
httpx==0.27.0
apscheduler==3.10.4
Pillow==10.3.0
python-dateutil==2.9.0
